resign with different account, delete old data

+++ ContentApp/BusinessLayer/Session/AIMSAccount.swift
@@ -136,12 +136,25 @@ extension AIMSAccount: AIMSAccountDelegate {
         createTicket()
         ProfileService.featchPersonalFilesID()
 
-        if oldAccountIdentifier != identifier {
+        if oldAccountIdentifier != identifier && oldAccountIdentifier != "" {
+            UserDefaults.standard.set(identifier, forKey: kActiveAccountIdentifier)
+            UserDefaults.standard.synchronize()
+            session.parameters.save(for: identifier)
+
             let path = DiskService.documentsDirectoryPath(for: oldAccountIdentifier)
             _ = DiskService.delete(itemAtPath: path)
 
             let listNodeDataAccessor = ListNodeDataAccessor()
             listNodeDataAccessor.removeAllNodes()
+
+            Keychain.delete(forKey: "\(oldAccountIdentifier)-\(String(describing: AlfrescoCredential.self))")
+            Keychain.delete(forKey: "\(oldAccountIdentifier)-\(String(describing: AlfrescoAuthSession.self))")
+            session.parameters.remove(for: oldAccountIdentifier)
         }
+
+        let notification = NSNotification.Name(rawValue: kReSignInNotification)
+        NotificationCenter.default.post(name: notification,
+                                        object: nil,
+                                        userInfo: nil)
     }
 }
